make project detail work
